+++ lua/jokers/bribery.lua
@@ -1,11 +1,12 @@
 local bribery = SMODS.Joker{
     atlas = 'joker',
     key = 'bribery',
-    pos = { x = 3, y = 1 }, soul_pos = { x = 0, y = 0 },
+    pos = { x = 4, y = 1 }, soul_pos = { x = 0, y = 0 },
     set_badges = function(self, card, badges) if (self.discovered) then
 		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Xtra)
 	end end,
     discovered = true,
+    blueprint_compat = false,
     rarity = 3,
     cost = 9,
     config = { extra = { cost = 50, used = true } },