+++ lua/jokers/flamenoble.lua
@@ -1,6 +1,6 @@
 SMODS.Joker {
     atlas = "joker",
-	key = "flame",
+	key = "flame_noble",
 	pos = { x = 1, y = 0 }, soul_pos = { x = 0, y = 0 },
     set_badges = function(self, card, badges) if (self.discovered) then
 		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Quake)