+++ lua/jokers/starla.lua
@@ -1,9 +1,7 @@
-if not next(SMODS.find_mod("ellejokers")) then return end
-
 SMODS.Joker{
     atlas = 'joker',
-    key = 'bubble',
-    pos = { x = 9, y = 1 }, soul_pos = { x = 0, y = 0 },
+    key = 'starla',
+    pos = { x = 9, y = 1 }, soul_pos = { x = 9, y = 2 },
     set_badges = function(self, card, badges) if (self.discovered) then
 		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Xtra)
 	end end,